Fix : Implement Error
CharacterBase에 추가된 Death 추가
float _> int형 변환

+++ Assets/Moon/Scripts/ICombat.cs
@@ -1,5 +1,5 @@
 
 public interface ICombat
 {
-    public void TakeDamage(EnemyType enemyType, float damageAmount);
+    public void TakeDamage(EnemyType enemyType, int damageAmount);
 }
